pyproject.toml
[project]
name = "algolabra"
version = "0.1.0"
description = ""
authors = [
    {name = "Emil Kivelä"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "flask (>=3.1.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "coverage (>=7.11.0,<8.0.0)"
]
